refactor : 위치 이동, 파라미터 추가
